refactor: extract ui components
