+++ packages/blocks/src/__internal__/rich-text/link-node/index.ts
@@ -1,5 +1,4 @@
 import './link-node';
-import './link-popover';
 
 export { LinkNodeComponent } from './link-node';
 export { createLink } from './create-link';
