--- packages/blocks/src/__internal__/rich-text/link-node/link-popover/index.ts
@@ -1,4 +1,0 @@
-import './button';
-import './link-popover';
-
-export { LinkPopover } from './link-popover';
