+++ packages/blocks/src/components/link-popover/index.ts
@@ -1,0 +1,5 @@
+import '../button';
+import './link-popover';
+
+export { showLinkPopover } from './create-link-popover';
+export { ALLOWED_SCHEMES } from './link-popover';
